2 usages of the NeoKey Trinkey
ITERATE --> maintain touch than click once the rename the selected file of the explorer / finder to increment it's name (useful when you want your files to be seen in a particuliar order)

Click twice and the added increment will be removed (attention, the device will juste erase the 3 first characters of the selected file, pay attention to what is currently selected)...

Click 4 times and the current window will close (ALT+F4)

COUNTER -->  maintain touch than click any times you want to increase the counter. When you'll let go of the touchpad, the selected fil on you explorer / finder will be renamed to start with "(the number of times you clicked)X_".

Usefull when you want to print x times a particuliar photo.

If you click the button without maintaining the touchpad,  the device will juste erase the 3 first characters of the selected file, pay attention to what is currently selected)...