Create OLED.py
Now a single device which makes the work of both NeoKey Trinkeys !
A feather M4 with the OLED SH1107 display and a NeoKey featherwing with 2 buttons !

Press  and maintain the left button then click the right button to iterate the current selected file of your explorer / finder.
Release to confirm.

Press  and maintain the right button then click the left button several times to add the number of times you want the currently selected file to be (printed in my case).
Release to confirm.

Press both buttons to remove what you've modified using the previous commands.